Multiple Linear regression and Backward elimination to optimal model.
Here the  question is to find the which company is making more profit, which city they are located. we had the data for 50 startup companies. and variables like R&D spend, Admin spend, Marketing and so on.
This file only shows the Data preprocessing, and model optimizing with back ward elimination assuming the significant level=0.05.